+++ docs/release-notes.rst
@@ -223,11 +223,15 @@ the registration system is not active.
   countries are official countries, when official and unofficial
   countries are distinguished at at least one event.
 
-* :file:`person.item.html` and :file:`page.html` have been modified
-  for internal changes in registration system code.
+* :file:`person.item.html`, :file:`person.scoreselect.html` and
+  :file:`page.html` have been modified for internal changes in
+  registration system code.
 
 * A bug has been fixed that produced internal errors when scaling down
   PNG photos with an alpha channel.
+
+* The menu of countries for entering scores no longer shows staff
+  countries or "None".
 
 Version 2018.02.0 (11 February 2018)
 ------------------------------------